+++ Assets/Tests/InputSystem/CoreTests_Actions.cs
@@ -9238,6 +9238,21 @@ partial class CoreTests
         Assert.That(action.GetBindingDisplayString(8), Is.EqualTo("Left Shift|Right Shift+A"));
     }
 
+    // https://issuetracker.unity3d.com/product/unity/issues/guid/UUM-141423
+    [Test]
+    [Category("Actions")]
+    public void Actions_WhenGettingDisplayTextForBindingsOnAction_CompositeIsIncludedWhenAtLeastOnePartMatchesBindingMask()
+    {
+        var action = new InputAction();
+
+        action.AddCompositeBinding("1DAxis")
+            .With("Negative", "<Keyboard>/a", groups: "Keyboard")
+            .With("Positive", "<Keyboard>/d", groups: "Keyboard");
+
+        Assert.That(action.GetBindingDisplayString(InputBinding.MaskByGroup("Keyboard")),
+            Is.EqualTo("A/D"));
+    }
+
     // https://fogbugz.unity3d.com/f/cases/1321175/
     [Test]
     [Category("Actions")]